fix waste code validator in registry
